+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+## [1.3.3] - 2020-06-19
+
+### Fixed
+
+- @react-corn/material-ui-lab meta prop is now mapping from title
+- @react-corn/material-ui-lab has now correct externals
+- Add Money field for main ui packages
+
 ## [1.3.2] - 2020-06-19
 
 - Rerelease all packages
@@ -74,7 +82,8 @@ Bad release
 
 ## [1.0.1] - 2020-06-04
 
-[unreleased]: https://github.com/paradoxxxzero/react-corn/compare/v1.3.2...HEAD
+[unreleased]: https://github.com/paradoxxxzero/react-corn/compare/v1.3.3...HEAD
+[1.3.3]: https://github.com/paradoxxxzero/react-corn/compare/v1.3.2...v1.3.3
 [1.3.2]: https://github.com/paradoxxxzero/react-corn/compare/v1.3.1...v1.3.2
 [1.3.1]: https://github.com/paradoxxxzero/react-corn/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/paradoxxxzero/react-corn/compare/v1.2.4...v1.3.0